+++ contracts/bridge/bin/addr/local-dvm.json
@@ -1,9 +1,9 @@
 {
   "DEPLOYER": "0x6Be02d1d3665660d22FF9624b7BE0551ee1Ac91b",
-  "FeeMarket": "0x89CaE076aaaA7A208c1d32960275844E89c8471E",
-  "BSCLightClient": "0xcBB1831E5f35648cD82cBC49F10752404770e8Fe",
-  "ChainMessageCommitter": "0x5Aae80231033c71Fe88827cAd4b0b401700327F8",
-  "LaneMessageCommitter": "0xCB9578aB2753ECB32CdE825CF14c53e7de7ae1a0",
-  "OutboundLane": "0xc57aa094A47627b3B7447A5b6E90df17b66fC03E",
-  "InboundLane": "0xdA1C5C54D96312a36c35dF8ACF285a52D894c677"
+  "FeeMarket": "0x45c69e7E21561B0CE3CcaE493Dc7808454be0afd",
+  "BSCLightClient": "0xAe6273952B921E6fCB8a3471ccD01FBD5d73230d",
+  "ChainMessageCommitter": "0xBff88cf042397d7BBE5D6401bd59136C51635A77",
+  "LaneMessageCommitter": "0x67B71dB1543750F172c5804875c29D84e7592c66",
+  "OutboundLane": "0x15c7EE8B39CD3eb95d842EF0c4c731335d849c28",
+  "InboundLane": "0x251e190c73F85218D00Ac3601f445db30FB49eA9"
 }
