+++ contracts/bridge/bin/addr/local-evm.json
@@ -1,7 +1,7 @@
 {
   "DEPLOYER": "0x71C7E24491eD49e4fc8fc20ef9B6F392CCfA654F",
-  "FeeMarket": "0x20B7856a6E28C019043fC3bbadAFc3330aF9Dc11",
-  "DarwiniaLightClient": "0xd1015f0212DFEa6842714F1EcFD6Ca8bddA54774",
-  "OutboundLane": "0x7f44939c482190cdA4c9f85a78de7cC51543A6CC",
-  "InboundLane": "0x1A3fabA8bc1b3D0bA3656e73a9FBf24746a37E1F"
+  "FeeMarket": "0x6C278db042eE12396F7c223D5861A0A7D9B94b32",
+  "DarwiniaLightClient": "0xb892a562A7c73Fd412Aa8c205bc4E138F1A12a73",
+  "OutboundLane": "0x979677Bd9FF4E062BA809bdBD36c4908F0135BA1",
+  "InboundLane": "0x3E05d8D6d8c8bAc9834f61642106f9eff707f8C8"
 }
