+++ contracts/bridge/e2e/message_storage_proof.js
@@ -1,7 +1,6 @@
 const { expect } = require("chai")
 const { waffle } = require("hardhat");
 const { BigNumber } = require("ethers");
-// const { solidity } = require("ethereum-waffle")
 const { bootstrap } = require("./helper/fixture")
 const chai = require("chai")
 const { solidity } = waffle;
